feat(FE): 리팩토링된 라우팅 및 인증 흐름 개선
- main.tsx에 있던 router를 App 컴포넌트로 옮겨 라우팅 구조 재구축
- 사용자 인증 상태를 관리하기 위해 useAuth 훅을 도입하고, LoginRequiredRoute 컴포넌트를 리팩토링하여 인증된 사용자만 접근할 수 있도록 설정함.
